+++ checkpoint/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 merging.
 - Add Patch for Pathways CPU ids.
 - Add sidecar MTC backend support
+- Add Colocated Python Sidecar MTC Checkpoint Manager
 
 ## [0.11.36] - 2026-04-14